Modify page with its recipe
Add the description of the dish, the ingredients, the steps to
complete the dish, and a picture of a cooked Spaghetti all'Assassina.

Except for the title (the name of the dish), subtopic heading
should start at level 2 (example: description, ingredients, and steps),
and descend sequentially if more subheading is needed.